+++ src/main/java/webserver/RequestHandler.java
@@ -1,5 +1,6 @@
 package webserver;
 
+import controller.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import utils.IOUtils;
@@ -9,10 +10,12 @@ import webserver.http.HttpResponse;
 import java.io.*;
 import java.net.Socket;
 
+import static utils.FileIoUtils.exists;
+
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
     private final Socket connection;
-    private final DispatcherServlet dispatcherServlet = new DispatcherServlet();
+    private final HandlerMapping handlerMapping = new HandlerMapping();
 
     public RequestHandler(Socket connectionSocket) {
         this.connection = connectionSocket;
@@ -24,12 +27,25 @@ public class RequestHandler implements Runnable {
 
         try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
             HttpRequest httpRequest = IOUtils.readRequest(new BufferedReader(new InputStreamReader(in)));
-            HttpResponse httpResponse = dispatcherServlet.process(httpRequest);
+            HttpResponse httpResponse = process(httpRequest);
             IOUtils.writeResponse(new DataOutputStream(out), httpResponse);
 
         } catch (IOException | NumberFormatException | NullPointerException e) {
             logger.error(e.getMessage());
         }
+    }
+
+    private Controller mapController(String uri) {
+        if (exists(uri)) {
+            return handlerMapping.getController("static");
+        }
+        return handlerMapping.getController("user");
+    }
+
+    private HttpResponse process(HttpRequest httpRequest) {
+        String uri = httpRequest.getUri();
+        Controller controller = mapController(uri);
+        return controller.mapRoute(httpRequest);
     }
 
 }
